refactor: delete sandbox resolution re-export
