+++ backend/threads/activity_pool_service.py
@@ -3,7 +3,7 @@
 from backend.threads.file_channel import get_file_channel_binding
 from backend.threads.pool import registry as _registry
 from backend.threads.pool.factory import create_agent_sync
-from backend.threads.sandbox_resolution import resolve_thread_sandbox
+from backend.threads.sandbox_resolution import resolve_thread_sandbox as _resolve_thread_sandbox
 from core.identity.agent_registry import get_or_create_agent_id
 
 
@@ -12,7 +12,7 @@ async def get_or_create_agent(*args, **kwargs):
     _registry.create_agent_sync = create_agent_sync
     _registry.get_or_create_agent_id = get_or_create_agent_id
     _registry.get_file_channel_binding = get_file_channel_binding
-    _registry.resolve_thread_sandbox = resolve_thread_sandbox
+    _registry.resolve_thread_sandbox = _resolve_thread_sandbox
     if "messaging_service" not in kwargs and app is not None:
         # @@@agent-pool-borrowed-runtime-bundle - thread runtime still needs a
         # chat-owned messaging service for chat_repos construction, but the
