+++ backend/web/core/dependencies.py
@@ -9,7 +9,8 @@ from backend.identity.auth.dependencies import _get_auth_service as _get_auth_se
 from backend.identity.auth.user_resolution import get_current_user as get_current_user
 from backend.identity.auth.user_resolution import get_current_user_id as get_current_user_id
 from backend.threads import convergence as thread_runtime_convergence
-from backend.threads.activity_pool_service import get_or_create_agent, resolve_thread_sandbox
+from backend.threads.activity_pool_service import get_or_create_agent
+from backend.threads.sandbox_resolution import resolve_thread_sandbox
 from sandbox.thread_context import set_current_thread_id
 
 inspect_owner_thread_runtime = thread_runtime_convergence.inspect_owner_thread_runtime
