+++ backend/web/routers/thread_files.py
@@ -7,7 +7,7 @@ from fastapi import APIRouter, Depends, File, HTTPException, Query, UploadFile
 from fastapi.responses import FileResponse
 
 from backend.threads import file_channel as file_channel_service
-from backend.threads.activity_pool_service import resolve_thread_sandbox
+from backend.threads.sandbox_resolution import resolve_thread_sandbox
 from backend.web.core.dependencies import get_app, verify_thread_owner
 from backend.web.utils.helpers import resolve_local_workspace_path
 from sandbox.thread_context import set_current_thread_id
